Move V1 tests from TokenUpdatePrecompileSuite to a new concurrent suite
